:building_construction: refactor: migrate to Turborepo + pnpm workspaces
Add packages/env and packages/utils as shared workspaces, lift the
nextra-theme-docs patch into the patches/ tree, and adjust apps/web
imports + workspace config to match the new layout.
