+++ README.zh-CN.md
@@ -1,11 +1,11 @@
 <div align="center">
   <a href="https://magic-resume.cn">
-    <img width="160" alt="Magic Resume Logo" src="./public/simple-logo.png">
+    <img width="160" alt="Magic Resume Logo" src="./apps/web/public/simple-logo.png">
   </a>
 
   <h1>Magic Resume</h1>
 
-  <p><strong>下一代 AI 简历平台，让招聘变得简单。</strong></p>
+  <p><strong>AI 原生简历平台 — 构建、分析、优化，并通过 AI 编程工具直接编辑简历。</strong></p>
 
 [English](./README.md) · **简体中文** · [官方网站][official-site] · [问题反馈][github-issues-link]
 
@@ -29,9 +29,12 @@
 - [✨ 功能特性](#-功能特性)
   - [制作：可视化模板自定义](#制作可视化模板自定义)
   - [分析：Lighthouse 风格报告](#分析lighthouse-风格报告)
-  - [优化：智能 JD 匹配](#优化智能-jd-匹配)
+  - [优化：AI 驱动的 JD 匹配](#优化ai-驱动的-jd-匹配)
+  - [AI Lab：面试练习与翻译](#ai-lab面试练习与翻译)
   - [隐私：本地优先的数据安全](#隐私本地优先的数据安全)
+- [🤖 MCP 集成](#-mcp-集成)
 - [🛳 私有化部署](#-私有化部署)
+  - [自托管模式（无需后端）](#自托管模式无需后端)
   - [使用 Vercel 部署](#使用-vercel-部署)
 - [📦 生态系统](#-生态系统)
 - [⌨️ 本地开发](#️-本地开发)
@@ -46,21 +49,23 @@
 
 <br/>
 
-![Banner](./public/magic-resume-preview.png)
+![Banner](./apps/web/public/magic-resume-preview.png)
 
 ## 👋🏻 快速开始
 
-**Magic Resume** 是一款现代化的 AI 驱动简历生成器，旨在帮助求职者轻松制作专业且极具影响力的简历。基于 Next.js 15 构建，它将优雅的用户界面与强大的 AI 能力相结合，助力您的求职之旅。
+**Magic Resume** 是一款基于 Turborepo monorepo 构建的现代化 AI 原生简历平台。它将实时可视化编辑器与多模型 AI 能力相结合——并更进一步，提供原生 **MCP 服务器**，让 AI 编程工具（Claude Code、Cursor、Windsurf）无需猜测数据结构，即可安全地读取和编辑简历。
 
 ## ✨ 功能特性
 
 ### 制作：可视化模板自定义
 
-使用我们直观的可视化编辑器，在几分钟内创建专业简历。
+使用直观的可视化编辑器，在几分钟内创建专业简历。
 
-- **实时预览**：输入时即可立即看到更改。
-- **灵活模板**：从专业且对 ATS 友好的模板中进行选择。
+- **实时预览**：输入即可立即看到变化。
+- **12 款专业模板**：对 ATS 友好，支持完整的颜色、字体和布局控制。
 - **丰富定制**：轻松调整颜色、字体（22+ 种样式）、间距和布局。
+- **版本历史**：保存快照并随时恢复任意历史版本。
+- **简历分享**：通过专属链接分享，支持查看者 / 评论者 / 编辑者权限控制。
 
 ### 分析：Lighthouse 风格报告
 
@@ -70,81 +75,141 @@
 - **详细分析**：深入了解关键词匹配、可行动性和可读性。
 - **可操作建议**：提供具体建议，让您的简历脱颖而出。
 
-### 优化：智能 JD 匹配
+### 优化：AI 驱动的 JD 匹配
 
 使用 AI 为特定职位描述量身定制简历。
 
-- **智能对齐**：AI 分析职位描述 (JD) 并建议内容优化。
-- **基于角色的建议**：针对特定行业和角色提供定制化建议。
+- **智能对齐**：AI 分析职位描述 (JD) 并建议针对性的内容优化。
+- **多模型支持**：兼容 OpenAI、Google Gemini、Anthropic Claude 及任意 OpenAI 兼容 API。
+
+### AI Lab：面试练习与翻译
+
+超越编辑，探索 AI 驱动的职业工具。
+
+- **面试练习**：模拟真实面试场景，AI 对回答给出即时反馈。
+- **简历翻译**：在保留格式的前提下，将简历翻译成任意语言。
 
 ### 隐私：本地优先的数据安全
 
 您的数据由您掌握。
 
-- **本地存储**：默认情况下，所有简历数据都存储在浏览器的本地。
-- **可选云端同步**：如果您选择，可以安全地跨设备同步数据。
-- **多格式导出**：将简历导出为高质量 PDF 或结构化 JSON。
+- **本地存储**：默认将所有简历数据存储在浏览器 IndexedDB 中，无需注册账号。
+- **可选云端同步**：选择性地跨设备安全同步数据。
+- **多格式导出**：导出为高质量 PDF 或结构化 JSON。
+
+---
+
+## 🤖 MCP 集成
+
+Magic Resume 提供原生 **Model Context Protocol (MCP) 服务器**（`@magic-resume/mcp`）——为 AI 编程工具打造的一等公民集成。
+
+配置一次个人访问令牌，您的 AI 助手即可安全地读取和修补简历，无需接触原始数据库：
+
+```bash
+npx -y @magic-resume/mcp config set --api-url "https://your-api.example.com/api" --pat "mr_pat_xxx"
+claude mcp add magic-resume -- npx -y @magic-resume/mcp mcp
+```
+
+**可用 MCP 工具：**
+
+| 工具 | 说明 |
+|---|---|
+| `list_resumes` | 列出账户中的所有简历 |
+| `get_resume` | 获取简历的完整内容 |
+| `get_resume_schema` | 获取用于验证的 JSON Schema |
+| `get_resume_editing_guide` | 获取 AI 可读的简历编辑指南 |
+| `preview_resume_patch` | 预览 JSON Patch 而不实际应用 |
+| `update_resume_content` | 应用 JSON Patch 更新简历内容 |
+
+MCP 服务器具备 schema 感知能力，采用基于 Patch 的操作方式——AI 工具进行精准修改而非全量覆盖。
 
 ---
 
 ## 🛠 技术栈
 
-- **框架**: [Next.js 15](https://nextjs.org/) (App Router)
-- **语言**: [TypeScript](https://www.typescriptlang.org/)
-- **AI / 大模型**: [LangChain](https://www.langchain.com/), [LangGraph](https://www.langchain.com/langgraph), [Google GenAI](https://ai.google.dev/), [Anthropic](https://www.anthropic.com/)
-- **身份认证**: [Clerk](https://clerk.com/)
-- **样式方案**: [Tailwind CSS 4](https://tailwindcss.com/)
-- **组件库**: [Radix UI](https://www.radix-ui.com/), [Lucide Icons](https://lucide.dev/)
-- **动画**: [Framer Motion](https://www.framer.com/motion/), [GSAP](https://gsap.com/)
-- **状态管理**: [Zustand](https://zustand-demo.pmnd.rs/)
-- **数据库/存储**: [IndexedDB](https://developer.mozilla.org/en-US/docs/Web/API/IndexedDB_API) (本地优先)
-- **富文本编辑器**: [Tiptap](https://tiptap.dev/), [Monaco Editor](https://microsoft.github.io/monaco-editor/)
-- **国际化**: [i18next](https://www.i18next.com/)
-- **分析工具**: [PostHog](https://posthog.com/)
+| 层次 | 技术 |
+|---|---|
+| Monorepo | [Turborepo](https://turbo.build/)、[pnpm](https://pnpm.io/) |
+| 框架 | [Next.js 15](https://nextjs.org/)（App Router） |
+| 语言 | [TypeScript](https://www.typescriptlang.org/) |
+| AI / 大模型 | [LangChain](https://www.langchain.com/)、[LangGraph](https://www.langchain.com/langgraph)、[Google GenAI](https://ai.google.dev/)、[Anthropic](https://www.anthropic.com/) |
+| 身份认证 | [Clerk](https://clerk.com/)（仅云端模式） |
+| 样式方案 | [Tailwind CSS 4](https://tailwindcss.com/) |
+| 组件库 | [Radix UI](https://www.radix-ui.com/)、[Lucide Icons](https://lucide.dev/) |
+| 动画 | [Framer Motion](https://www.framer.com/motion/)、[GSAP](https://gsap.com/) |
+| 状态管理 | [Zustand](https://zustand-demo.pmnd.rs/) |
+| 本地存储 | [IndexedDB](https://developer.mozilla.org/en-US/docs/Web/API/IndexedDB_API) |
+| 富文本编辑器 | [Tiptap](https://tiptap.dev/)、[Monaco Editor](https://microsoft.github.io/monaco-editor/) |
+| MCP | [@modelcontextprotocol/sdk](https://modelcontextprotocol.io/) |
+| Schema | [Zod](https://zod.dev/) |
+| 国际化 | [i18next](https://www.i18next.com/) |
 
 ---
 
 ## 🛳 私有化部署
 
-在几分钟内部署您自己的 Magic Resume 实例。
+### 自托管模式（无需后端）
+
+Magic Resume 可以完全在浏览器中运行，无需后端、数据库或账号注册。设置 `NEXT_PUBLIC_APP_MODE=self-hosted`（或留空），所有数据将持久化在 IndexedDB 中。
+
+```bash
+git clone https://github.com/LinMoQC/Magic-Resume.git
+cd Magic-Resume
+cp apps/web/.env.example apps/web/.env.local
+# 编辑 .env.local：设置 NEXT_PUBLIC_APP_MODE=self-hosted
+pnpm install
+pnpm run dev
+```
 
 ### 使用 Vercel 部署
 
-点击下方按钮部署到 Vercel：
+如需完整的云端体验（认证、云同步、分享功能），点击下方按钮一键部署：
 
 [![Deploy with Vercel](https://vercel.com/button)](https://vercel.com/new/clone?repository-url=https%3A%2F%2Fgithub.com%2FLinMoQC%2FMagic-Resume)
 
 > [!TIP]
->
-> 请记得在 Vercel 控制面板中配置您的 `NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY` 和 `CLERK_SECRET_KEY`。
+> 云端模式需在 Vercel 控制面板中配置 `NEXT_PUBLIC_APP_MODE=cloud`、`NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY` 和 `CLERK_SECRET_KEY`。
 
 ---
 
 ## 📦 生态系统
 
-- **Magic Resume Core**：主简历生成器和 AI 引擎。
-- **i18n Scanner**：自定义工具，确保完整的本地化覆盖。
+本仓库是一个 Turborepo monorepo，核心包如下：
+
+| 包 | 说明 |
+|---|---|
+| `apps/web` | Next.js 前端——编辑器、仪表盘、AI Lab |
+| `packages/mcp` | `@magic-resume/mcp`——stdio MCP 服务器和 CLI |
+| `packages/resume-schema` | 共享 Zod Schema、类型定义和示例数据 |
+| `packages/resume-templates` | 模板 DSL、渲染器和注册表 |
 
 ---
 
 ## ⌨️ 本地开发
 
-克隆仓库并启动开发服务器：
-
 ```bash
-$ git clone https://github.com/LinMoQC/Magic-Resume.git
-$ cd Magic-Resume
-$ npm install
-$ npm run dev
-```
+# 安装依赖
+pnpm install
 
-更多详情，请查看我们的 [开发指南](./docs/development.md)（即将上线）。
+# 启动所有工作区
+pnpm run dev
+
+# 或针对单个工作区
+pnpm --filter @magic-resume/web dev
+pnpm --filter @magic-resume/mcp build
+
+# 代码检查与测试
+pnpm run lint
+pnpm run test
+```
 
 ---
 
 ## 🤝 参与贡献
 
 欢迎贡献！请随时提交 Pull Request。
+
+如需报告 Bug 或提交功能建议，请[提交 Issue][github-issues-link]。
 
 <a href="https://github.com/LinMoQC/Magic-Resume/graphs/contributors">
   <img src="https://contrib.rocks/image?repo=LinMoQC/Magic-Resume" alt="contributors" />
@@ -176,6 +241,3 @@ Copyright © 2026 [Magic Resume Team](https://github.com/LinMoQC). <br />
 [github-issues-shield]: https://img.shields.io/github/issues/LinMoQC/Magic-Resume?color=ff80eb&labelColor=black&style=flat-square
 [github-license-shield]: https://img.shields.io/badge/license-MIT-white?labelColor=black&style=flat-square
 [github-license-link]: https://github.com/LinMoQC/Magic-Resume/blob/master/LICENSE
-[github-release-link]: https://github.com/LinMoQC/Magic-Resume/releases
-[github-release-shield]: https://img.shields.io/github/v/release/LinMoQC/Magic-Resume?color=369eff&labelColor=black&logo=github&style=flat-square
-[image-star]: https://github.com/user-attachments/assets/3216e25b-186f-4a54-9cb4-2f124aec0471
