+++ apps/web/package.json
@@ -25,7 +25,9 @@
     "@langchain/google-genai": "^0.2.10",
     "@langchain/langgraph": "^0.3.3",
     "@langchain/tavily": "^0.1.3",
+    "@magic-resume/env": "workspace:*",
     "@magic-resume/resume-templates": "workspace:*",
+    "@magic-resume/utils": "workspace:*",
     "@microlink/react-json-view": "^1.26.2",
     "@monaco-editor/react": "^4.7.0",
     "@radix-ui/react-avatar": "^1.1.10",
