+++ apps/web/src/lib/utils/utils.ts
@@ -1,35 +1,8 @@
-import { type ClassValue, clsx } from "clsx"
-import { twMerge } from "tailwind-merge"
 import { MagicDebugger } from "./debuggger";
 import html2canvas from "html2canvas";
+export { cn } from '@magic-resume/utils';
+export { formatTime, formatRelativeTime } from '@magic-resume/utils';
 
-export function cn(...inputs: ClassValue[]) {
-  return twMerge(clsx(inputs))
-}
-
-// 时间格式化
-export function formatTime(ts: number) {
-  const date = new Date(ts);
-  const now = new Date();
-  const diff = now.getTime() - date.getTime();
-
-  const diffSeconds = Math.floor(diff / 1000);
-  const diffMinutes = Math.floor(diffSeconds / 60);
-  const diffHours = Math.floor(diffMinutes / 60);
-  const diffDays = Math.floor(diffHours / 24);
-
-  if (diffSeconds < 60) {
-    return `${diffSeconds}s ago`;
-  } else if (diffMinutes < 60) {
-    return `${diffMinutes}m ago`;
-  } else if (diffHours < 24) {
-    return `${diffHours}h ago`;
-  } else if (diffDays < 7) {
-    return `${diffDays}d ago`;
-  } else {
-    return date.toLocaleDateString();
-  }
-} 
 
 // 简历快照
 export function generateSnapshot(options?: {
@@ -191,22 +164,6 @@ function oklchToRgb(l: number, c: number, h: number){
   return [r * 255, g * 255, bl * 255];
 }
 
-export function formatRelativeTime(date: string) {
-  const now = new Date();
-  const targetDate = new Date(date);
-  const diffInSeconds = Math.floor((now.getTime() - targetDate.getTime()) / 1000);
-  
-  const minutes = Math.floor(diffInSeconds / 60);
-  const hours = Math.floor(minutes / 60);
-  const days = Math.floor(hours / 24);
-  
-  if (diffInSeconds < 60) return `${diffInSeconds}s ago`;
-  if (minutes < 60) return `${minutes}m ago`;
-  if (hours < 24) return `${hours}h ago`;
-  if (days < 7) return `${days}d ago`;
-  
-  return targetDate.toLocaleDateString();
-}
 
 /**
  * 客户端安全的时间生成函数
