+++ patches/nextra-theme-docs@4.3.0.patch
@@ -1,0 +1,167 @@
+diff --git a/dist/layout.js b/dist/layout.js
+index 57a511faacf75820f822be305d66c61494ca998a..f1ea6f6e8afe693a1f387b687eb73d1578ee6bdf 100644
+--- a/dist/layout.js
++++ b/dist/layout.js
+@@ -1,125 +1,45 @@
+ import { jsx, jsxs } from "react/jsx-runtime";
+-import { c as _c } from "react-compiler-runtime";
+ import { ThemeProvider } from "next-themes";
+ import { SkipNavLink } from "nextra/components";
+ import { z } from "zod";
+ import { MobileNav } from "./components/sidebar";
+ import { LayoutPropsSchema } from "./schemas";
+ import { ConfigProvider, ThemeConfigProvider } from "./stores";
+-const Layout = (t0) => {
+-  const $ = _c(26);
+-  let children;
+-  let themeConfig;
+-  if ($[0] !== t0) {
+-    ({
+-      children,
+-      ...themeConfig
+-    } = t0);
+-    $[0] = t0;
+-    $[1] = children;
+-    $[2] = themeConfig;
+-  } else {
+-    children = $[1];
+-    themeConfig = $[2];
+-  }
+-  let data;
+-  if ($[3] !== themeConfig) {
+-    const {
+-      data: t12,
+-      error
+-    } = LayoutPropsSchema.safeParse(themeConfig);
+-    data = t12;
+-    if (error) {
+-      throw z.prettifyError(error);
+-    }
+-    $[3] = themeConfig;
+-    $[4] = data;
+-  } else {
+-    data = $[4];
+-  }
+-  let banner;
+-  let footer;
+-  let navbar;
+-  let nextThemes;
+-  let pageMap;
+-  let rest;
+-  if ($[5] !== data) {
+-    ({
+-      footer,
+-      navbar,
+-      pageMap,
+-      nextThemes,
+-      banner,
+-      ...rest
+-    } = data);
+-    $[5] = data;
+-    $[6] = banner;
+-    $[7] = footer;
+-    $[8] = navbar;
+-    $[9] = nextThemes;
+-    $[10] = pageMap;
+-    $[11] = rest;
+-  } else {
+-    banner = $[6];
+-    footer = $[7];
+-    navbar = $[8];
+-    nextThemes = $[9];
+-    pageMap = $[10];
+-    rest = $[11];
+-  }
+-  let t1;
+-  if ($[12] === Symbol.for("react.memo_cache_sentinel")) {
+-    t1 = /* @__PURE__ */ jsx(SkipNavLink, {});
+-    $[12] = t1;
+-  } else {
+-    t1 = $[12];
+-  }
+-  let t2;
+-  if ($[13] === Symbol.for("react.memo_cache_sentinel")) {
+-    t2 = /* @__PURE__ */ jsx(MobileNav, {});
+-    $[13] = t2;
+-  } else {
+-    t2 = $[13];
+-  }
+-  let t3;
+-  if ($[14] !== children || $[15] !== footer || $[16] !== navbar || $[17] !== pageMap) {
+-    t3 = /* @__PURE__ */ jsxs(ConfigProvider, { pageMap, navbar, footer, children: [
+-      t2,
+-      children
+-    ] });
+-    $[14] = children;
+-    $[15] = footer;
+-    $[16] = navbar;
+-    $[17] = pageMap;
+-    $[18] = t3;
+-  } else {
+-    t3 = $[18];
+-  }
+-  let t4;
+-  if ($[19] !== banner || $[20] !== nextThemes || $[21] !== t3) {
+-    t4 = /* @__PURE__ */ jsxs(ThemeProvider, { ...nextThemes, children: [
+-      t1,
+-      banner,
+-      t3
+-    ] });
+-    $[19] = banner;
+-    $[20] = nextThemes;
+-    $[21] = t3;
+-    $[22] = t4;
+-  } else {
+-    t4 = $[22];
+-  }
+-  let t5;
+-  if ($[23] !== rest || $[24] !== t4) {
+-    t5 = /* @__PURE__ */ jsx(ThemeConfigProvider, { value: rest, children: t4 });
+-    $[23] = rest;
+-    $[24] = t4;
+-    $[25] = t5;
+-  } else {
+-    t5 = $[25];
+-  }
+-  return t5;
+-};
+-export {
+-  Layout
++
++const Layout = (props) => {
++  const { data, error } = LayoutPropsSchema.safeParse(props);
++  if (error) {
++    throw z.prettifyError(error);
++  }
++  const {
++    footer,
++    navbar,
++    pageMap,
++    nextThemes,
++    banner,
++    children,
++    ...rest
++  } = data;
++
++  return jsx(ThemeConfigProvider, {
++    value: rest,
++    children: jsxs(ThemeProvider, {
++      ...nextThemes,
++      children: [
++        jsx(SkipNavLink, {}),
++        banner,
++        jsxs(ConfigProvider, {
++          pageMap,
++          navbar,
++          footer,
++          children: [
++            jsx(MobileNav, {}),
++            children,
++          ],
++        }),
++      ],
++    }),
++  });
+ };
++
++export { Layout };
